+++ lognflow/plt_utils.py
@@ -1,3 +1,4 @@
+from .printprogress import printprogress
 from mpl_toolkits.axes_grid1 import make_axes_locatable
 import numpy as np
 import matplotlib.pyplot as plt
@@ -41,10 +42,12 @@ def pltfig_to_numpy(fig):
     w,h = fig.canvas.get_width_height()
     buf = np.frombuffer(fig.canvas.tostring_argb(), dtype=np.ubyte)
     buf.shape = (w, h, 4)
-    buf = np.roll (buf, 3, axis = 2)
-    return buf
+    return buf.sum(2)
 
-def numbers_as_images_3D(dataset_shape: tuple, fontsize, verbose = False):
+def numbers_as_images_3D(data3D_shape: tuple[int, int, int],
+                         fontsize: int, 
+                         text_loc: tuple[int, int] = None,
+                         verbose: bool = True):
     """ Numbers3D
     This function generates a 4D dataset of images with shape
     (n_x, n_r, n_c) where in each image the value "x" is written as a text
@@ -61,8 +64,12 @@ def numbers_as_images_3D(dataset_shape: tuple, fontsize, verbose = False):
     output 3D array.
     
     """
-    n_x, n_r, n_c = dataset_shape
-    dataset = np.zeros(dataset_shape)    
+    n_x, n_r, n_c = data3D_shape
+    
+    if text_loc is None:
+        text_loc = (n_r//2 - fontsize, n_c//2 - fontsize)
+    
+    dataset = np.zeros(data3D_shape)    
     txt_width = int(np.log(n_x)/np.log(n_x)) + 1
     number_text_base = '{ind_x:0{width}}}'
     if(verbose):
@@ -71,22 +78,23 @@ def numbers_as_images_3D(dataset_shape: tuple, fontsize, verbose = False):
         mat = np.ones((n_r, n_c))
         number_text = number_text_base.format(ind_x = ind_x, 
                                               width = txt_width)
-        fig = plt.figure(figsize = (1, 1))
+        fig = plt.figure(figsize = (n_rr, n_cc), dpi = n_rc)
         ax = fig.add_subplot(111)
         ax.imshow(mat, cmap = 'gray', vmin = 0, vmax = 1)
-        ax.text(mat.shape[0]//2 - fontsize, mat.shape[1]//2 ,
+        ax.text(text_loc[0], text_loc[1],
                 number_text, fontsize = fontsize)
         ax.axis('off')
         buf = pltfig_to_numpy(fig)
         plt.close()
-        buf2 = buf[::buf.shape[0]//n_r, ::buf.shape[1]//n_c, :3].mean(2)
-        buf2 = buf2[:n_r, :n_c]
-        dataset[ind_x] = buf2.copy()
+        dataset[ind_x] = buf.copy()
         if(verbose):
             pBar()
     return dataset
 
-def numbers_as_images_4D(dataset_shape: tuple, fontsize, verbose = False):
+def numbers_as_images_4D(data4D_shape: tuple[int, int, int, int],
+                         fontsize: int, 
+                         text_loc: tuple[int, int] = None,
+                         verbose: bool = True):
     """ Numbers4D
     This function generates a 4D dataset of images with shape
     (n_x, n_y, n_r, n_c) where in each image the value "x, y" is written as a text
@@ -102,31 +110,38 @@ def numbers_as_images_4D(dataset_shape: tuple, fontsize, verbose = False):
     You can provide a logs_root, log_dir or simply select a directory to save the
     output 4D array.
     
+    :param text__loc:
+        text_loc should be a tuple of the location of bottom left corner of the
+        text in the image.
+    
     """
-    n_x, n_y, n_r, n_c = dataset_shape
+    n_x, n_y, n_r, n_c = data4D_shape
+
+    if text_loc is None:
+        text_loc = (n_r//2 - fontsize, n_c//2 - fontsize)
+    
     dataset = np.zeros((n_x, n_y, n_r, n_c))    
     txt_width = int(np.log(np.maximum(n_x, n_y))
-                    /np.log(np.maximum(n_x, n_y))) + 1
+                    / np.log(np.maximum(n_x, n_y))) + 1
     number_text_base = '{ind_x:0{width}}, {ind_y:0{width}}'
     if(verbose):
         pBar = printprogress(n_x * n_y)
     for ind_x in range(n_x):
         for ind_y in range(n_y):
             mat = np.ones((n_r, n_c))
-            number_text = number_text_base.format(ind_x = ind_x, 
-                                                  ind_y = ind_y,
-                                                  width = txt_width)
-            fig = plt.figure(figsize = (1, 1))
+            number_text = number_text_base.format(
+                ind_x = ind_x, ind_y = ind_y, width = txt_width)
+            n_rc = np.minimum(n_r, n_c)
+            n_rr = n_r / n_rc
+            n_cc = n_c / n_rc
+            fig = plt.figure(figsize = (n_rr, n_cc), dpi = n_rc)
             ax = fig.add_subplot(111)
             ax.imshow(mat, cmap = 'gray', vmin = 0, vmax = 1)
-            ax.text(mat.shape[0]//2 - fontsize, mat.shape[1]//2 ,
-                    number_text, fontsize = fontsize)
+            ax.text(text_loc[0], text_loc[1], number_text, fontsize = fontsize)
             ax.axis('off')
             buf = pltfig_to_numpy(fig)
             plt.close()
-            buf2 = buf[::buf.shape[0]//n_r, ::buf.shape[1]//n_c, :3].mean(2)
-            buf2 = buf2[:n_r, :n_c]
-            dataset[ind_x, ind_y] = buf2.copy()
+            dataset[ind_x, ind_y] = buf.copy()
             if(verbose):
                 pBar()
     return dataset